requirements.txt
Flask==3.0.3
Werkzeug==3.0.1
pillow==10.3.0
pytesseract==0.3.10
layoutparser[layoutmodels]==0.3.4
opencv-python==4.9.0.80
pdf2image==1.17.0
torch>=2.0.0
torchvision>=0.15.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
numpy>=1.24.0
scipy>=1.11.0
typing-extensions>=4.7.0  # Required for Protocol support
reportlab==4.1.0  # Required for PDF generation in tests 